ESM + TypeScript
I'd like to get AirPlay working between my iPad and OBS on my Mac. I found another library that might appear to have this functionality, but I'm not sure it's designed for macOS. I'm more of a JS and or Node guy myself, so this project seemed like a nice place to start if I were to set up my own receiver.

Was this in a working state when you last worked on it? I'd be interested in helping maintain further updates here too if you'd like.

Here's the project I was hoping to figure out recreating with Node instead, maybe. Do you have any insight on how I could pull this off? Thank you!

https://github.com/mika314/obs-airplay